feat: make numbers formatting locale aware
Use the current OS locale to format numbers.
The only place where numbers formatting is left untouched is for the "weight" as it makes sense in the current state.

In those changes, the text field containing the bill amount is now displayng the full raw number for correctness, and number format changes are not compatible anyway.